fix(db): move db path to another file (couldn't find module drizzle.config)
