+++ apps/api/drizzle.config.ts
@@ -1,9 +1,5 @@
-import path from "node:path";
-import { getAppDataPath } from "@local-sql/utils/get-appdata-path";
 import { defineConfig } from "drizzle-kit";
-
-export const DB_PATH = path.join(getAppDataPath(), "data.db");
-export const DB_PATH_FILE = `file:${DB_PATH}`;
+import { DB_PATH_FILE } from "./src/db/db-path";
 
 export default defineConfig({
   dialect: "sqlite",
