+++ apps/api/src/db/db-path.ts
@@ -1,6 +1,9 @@
 import fs from "node:fs/promises";
 import path from "node:path";
-import { DB_PATH } from "drizzle.config";
+import { getAppDataPath } from "@local-sql/utils/get-appdata-path";
+
+export const DB_PATH = path.join(getAppDataPath(), "data.db");
+export const DB_PATH_FILE = `file:${DB_PATH}`;
 
 let isDirectoryCreated = false;
